% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wsl_pseudoabsences.R
\name{plot.wsl.pseudoabsences}
\alias{plot.wsl.pseudoabsences}
\title{A simple plotting function for objects of class 'wsl.pseudoabsences'}
\usage{
plot.wsl.pseudoabsences(object)
}
\arguments{
\item{object}{An object of class wsl.pseudoabsences}
}
\value{
Presences (red crosses) and pseudoabsences (grey transparent points) are plotted in geographic space. If
the wsl.pseudoabsences object contains information on a template raster file (see documentation of the
wsl.samplePseuAbs function), this template will be plotted in the background.
}
\description{
This function allos quickly visualizing the different thinning and sampling
strategies determine the spatial patterns of presences and pseudoabsences.
}
