% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_pseudoabsences_class.R
\docType{class}
\name{wsl.pseudoabsences-class}
\alias{wsl.pseudoabsences-class}
\alias{wsl.pseudoabsences}
\title{An S4 class to store pseudoabsence data}
\description{
Information on coordinates, presence/pseudoabsence, and environmental
conditions is organized in seperate slots. Moreover, the slot meta
keeps track of meta information.
}
\section{Slots}{

\describe{
\item{\code{meta}}{a list with meta information}

\item{\code{pa}}{a vector with 1 encoding presences and 0 encoding pseudoabsences}

\item{\code{env_vars}}{a data.frame containing extractions of the env.stack at the points xy}

\item{\code{xy}}{two-column matrix with coordinates of the presence points
and the sampled pseudoabsence points. The projection is the same as the one of the env.stack supplied}

\item{\code{call}}{the call made to the wsl.samplePseuAbs function}
}}

\author{
Philipp Brun
}
